+++ src/apis/companies/mocks/myCompaniesMockHandler.ts
@@ -1,5 +1,5 @@
 import { APIPath } from '@/apis/apiPath';
-import { companyList } from '@/pages/myPage/employer/EmployerMyPage.mock';
+import { companyList } from './myCompanies.mock';
 import { http, HttpResponse } from 'msw';
 
 export const myCompaniesMockHandler = [http.get(APIPath.getMyCompanies, () => HttpResponse.json(companyList))];
