+++ src/features/applicants/ApplicantList/ContractModal/ModalText.styles.ts
