+++ src/features/applicants/ApplicantList/ContractModal/ModalText.tsx
