+++ src/pages/applicants/Applicants.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 import Applicants from './Applicants';
-import { recruitment, applicantList } from './Applicants.mock';
+import { recruitment, applicantList } from '../../apis/applicants/mocks/applicants.mock';
 
 const meta: Meta<typeof Applicants> = {
   title: 'pages/Applicants',
